Add Homepage classes for Wales/NI.
