+++ classes/Homepage/Base.php
@@ -18,7 +18,7 @@ abstract class Base {
 
         $data['regional'] = $this->getRegionalList();
         $data['popular_searches'] = $common->getPopularSearches();
-        $data['featured'] = $this->getEditorialContent();
+        $data['featured'] = $this->getEditorialContent($data);
         $data['divisions'] = $this->getRecentDivisions();
         $data['search_box'] = $this->getSearchBox($data);
 
@@ -26,7 +26,7 @@ abstract class Base {
     }
 
     abstract protected function getSearchBox(array $data): \MySociety\TheyWorkForYou\Search\SearchBox;
-    abstract protected function getEditorialContent();
+    abstract protected function getEditorialContent(array &$data);
 
     protected function getRegionalList() {
         return null;
@@ -48,6 +48,15 @@ abstract class Base {
             $more_url = new \MySociety\TheyWorkForYou\Url($recent[1]);
             if ($recent[0] == 'recent_pbc_debates') {
                 $content = [ 'data' => $instance->display($recent[0], ['num' => 5], 'none') ];
+            } elseif ($recent[1] == 'senedddebatesfront' || $recent[1] == 'nidebatesfront') {
+                $content = $instance->display($recent[0], ['days' => 30, 'num' => 6], 'none');
+                # XXX Bit hacky, for now
+                foreach ($content['data'] as $d) {
+                    $d['more_url'] = $more_url->generate();
+                    $d['desc'] = '';
+                    $recent_content[] = $d;
+                }
+                $content = [];
             } else {
                 $content = $instance->display($recent[0], ['days' => 7, 'num' => 1], 'none');
                 if (isset($content['data']) && count($content['data'])) {
